aws/sera-api/requirements.txt
mysql-connector-python
pandas
numpy